Update build for new project; allow testing config files elsewhere

+++ v9t9/v9t9-java/v9t9-common/src/v9t9/common/client/ISoundHandler.java
@@ -22,7 +22,7 @@ public interface ISoundHandler {
 			"PlaySound", new Boolean(true));
 	SettingSchema settingSoundVolume = new SettingSchema(
 			ISettingsHandler.WORKSPACE,
-			"SoundVolume", new Integer(10));
+			"SoundVolume", new Integer(5));
 
 
 	/**
